Fix: Socket::read() not returning a packet when $force is true
In certain cases, Socket::read() may not return a packet, even though
$force is set to true.

This happens when packets are relatively large (typically > 1500 bytes,
headers included). Xat's server's IP stack fragments the packet in two
or more packets.

If we're lucky enough, our kernel will already have reassembled the
fragmented packets when we're calling read(). Even though this is what
happens most of the time, we may call read() between two packets.
The script will therefore read the first parts of a fragmented packet.

The workaround is to wait for the socket's buffer to have a NULL byte
(which is used to separate packets in xat's protocol)